Move ObjectField and inflate() to own files + allow any prop to MyTextField
So they can be reused outside ExperimentConfigForm
